Prefer static builds for now
We rather statically link the from the external projects, the reason
is that oxen-core and lokinet for example links to different versions
of librairies due to different release timeline and many
complications. Ideally, in the future, we could get everything
as to link against shared libraries (with an ebuild for each librairies).